pyproject.toml
[project]
name = "learn-langchain-udemy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "langchain>=1.2.3",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-openai>=1.1.7",
    "python-dotenv>=1.2.1",
]
